+++ angel_system/data/common/kwcoco_utils.py
@@ -1,4 +1,3 @@
-import re
 import os
 import cv2
 import kwcoco
@@ -9,156 +8,10 @@ import ubelt as ub
 import pandas as pd
 
 from pathlib import Path
-from typing import List
 
 from detectron2.data.detection_utils import read_image
-from utils.data.dataloaders.structures import Activity
-
-RE_FILENAME_TIME = re.compile(r"frame_(?P<frame>\d+)_(?P<ts>\d+(?:_|.)\d+).(?P<ext>\w+)")
-def time_from_name(fname):
-    """
-    Extract the float timestamp from the filename.
-
-    :param fname: Filename of an image in the format
-        frame_<frame number>_<seconds>_<nanoseconds>.<extension>
-
-    :return: timestamp (float) in seconds
-    """
-    fname = os.path.basename(fname)
-    match = RE_FILENAME_TIME.match(fname)
-    time = match.group('ts')
-    if '_' in time:
-        time = time.split('_')
-        time = float(time[0]) + (float(time[1]) * 1e-9)
-    elif '.' in time:
-        time = float(time)
-
-    frame = match.group('frame')
-    return int(frame), time
-
-def activities_from_dive_csv(filepath: str) -> List[Activity]:
-    """
-    Load from a DIVE output CSV file a sequence of ground truth activity
-    annotations.
-
-    Class labels are converted to lower-case and stripped of any extraneous
-    whitespace.
-
-    :param filepath: Filesystem path to the CSV file.
-    :return: List of loaded activity annotations.
-    """
-    print(f"Loading ground truth activities from: {filepath}")
-    df = pd.read_csv(filepath)
-    # There may be additional metadata rows. Filter out rows whose first column
-    # value starts with a `#`.
-    df = df[df[df.keys()[0]].str.contains('^[^#]')]
-    # Create a mapping of detection/track ID to the activity annotation
-    id_to_activity: Dict[int, Activity] = {}
-    for row in df.iterrows():
-        i, s = row
-        a_id = int(s[0])
-        frame, time = time_from_name(s[1])
-        if a_id not in id_to_activity:
-            id_to_activity[a_id] = Activity(
-                s[9].lower().strip(), # class label
-                time, # start
-                np.inf, # end 
-                frame, # start frame
-                np.inf, # end frame
-                1.0, # conf
-            )
-        else:
-            # There's a struct in there, update it.
-            a = id_to_activity[a_id]
-            # Activity should not already have an end time assigned.
-            assert a.end is np.inf, (
-                f"More than 2 entries observed for activity track ID {a_id}."
-            )
-            id_to_activity[a_id] = Activity(
-                a.class_label,
-                a.start,
-                time,
-                a.start_frame,
-                frame,
-                a.conf,
-            )
-    # Assert that all activities have been assigned an associated end time.
-    assert np.inf not in {a.end for a in id_to_activity.values()}, (
-        f"Some activities in source CSV do not have corresponding end time "
-        f"entries: {filepath}"
-    )
-    return list(id_to_activity.values())
-
-def Re_order(image_list, image_number):
-    img_id_list = []
-    for img in image_list:
-        img_id, ts = time_from_name(img)
-        img_id_list.append(img_id)
-    img_id_arr = np.array(img_id_list)
-    s = np.argsort(img_id_arr)
-    new_list = []
-    for i in range(image_number):
-        idx = s[i]
-        new_list.append(image_list[idx])
-    return new_list
-
-def find_matching_gt_activity(gt_activity, fn):
-    fn = os.path.basename(fn)
-    frame, time = time_from_name(fn)
-
-    """
-    gt_activity = {
-        sub_step_str: [{
-            'start': 123456,
-            'end': 657899
-        }]
-    }
-    """
-
-    matching_gt = {}
-    for sub_step_str, times in gt_activity.items():
-        for gt_time in times:
-            if gt_time['start'] <= time <= gt_time['end']:
-                return sub_step_str
-            
-    return 'None'
 
 
-    
-def coffee_activity_data_loader(video='all_activities_11'):
-    import pandas as pd
-    # Load ground truth activity 
-    #root_dir = '/media/hannah.defazio/Padlock_DT/Data/notpublic/PTG/Coffee/coffee_labels'
-    root_dir = '/Padlock_DT/Coffee/coffee_labels'
-    gt_dir = f'{root_dir}/Labels'
-
-    gt_activity_fn = f'{gt_dir}/{video}.feather'
-    print(f"Loaded ground truth from {gt_activity_fn}")
-
-    gt = pd.read_feather(gt_activity_fn) # Keys: class, start_frame,  end_frame, exploded_ros_bag_path
-    gt_activity = {}
-    for i, row in gt.iterrows():
-        class_label = row["class"].lower().strip().strip('.')
-        if class_label not in gt_activity.keys():
-            gt_activity[class_label] = []
-        start_frame, start_time = time_from_name(row["start_frame"])
-        end_frame, end_time = time_from_name(row["end_frame"])
-        gt_activity[class_label].append({
-            'start': start_frame,#start_time,
-            'end': end_frame#end_time
-        })
-    print(f"Loaded ground truth from {gt_activity_fn}")
-
-
-    """
-    gt_activity = {
-        sub_step_str: [{
-            'start': 123456,
-            'end': 657899
-        }]
-    }
-    """
-    return gt_activity
 
 
 # Save
@@ -233,6 +86,27 @@ def preds_to_kwcoco(metadata, preds, save_dir, save_fn='result-with-contact.msco
     print(f'Saved predictions to {dset.fpath}')
 
     return dset
+
+def print_class_freq(dset):
+    freq_per_class = dset.category_annotation_frequency()
+    stats = []
+
+    for cat in dset.cats.values():
+        freq = freq_per_class[cat['name']]
+        class_ = {
+            'id': cat['id'],
+            'name': cat['name'],
+            #'instances_count': freq,
+            #'def': '',
+            #'synonyms': [],
+            #'image_count': freq,
+            #'frequency': '',
+            #'synset': ''
+        }
+
+        stats.append(class_)
+
+    print(f'MC50_CATEGORIES = {stats}')
 
 def visualize_kwcoco(dset=None, save_dir=''):
     import matplotlib.pyplot as plt
@@ -316,20 +190,138 @@ def visualize_kwcoco(dset=None, save_dir=''):
         plt.close(fig) # needed to remove the plot because savefig doesn't clear it
     plt.close('all')
 
+def filter_kwcoco(dset, split):
+    experiment_name = 'm2_all_data_cleaned_fixed_with_steps'
+    stage = 'stage2'
+
+    print('Experiment: ', experiment_name)
+    print('Stage: ', stage)
+
+    if type(dset) == str:
+        print(f'Loaded dset from file: {dset}')
+        dset = kwcoco.CocoDataset(dset)
+        print(dset)
+
+    # Remove in-between categories
+    remove_cats = []
+    for cat_id in dset.cats:
+        cat_name = dset.cats[cat_id]['name']
+        if '.5)' in cat_name or '(before)' in cat_name or '(finished)' in cat_name:
+            remove_cats.append(cat_id)        
+            
+    print(f'removing cat ids: {remove_cats}')
+    dset.remove_categories(remove_cats)
+
+    # Remove images with these 
+    gid_to_aids = dset.index.gid_to_aids
+    gids = ub.argsort(ub.map_vals(len, gid_to_aids))
+
+    remove_images = []
+    remove_anns = []
+    for gid in sorted(gids):
+        im = dset.imgs[gid]
+        
+        fn = im['file_name'].split('/')[-1]
+        gt = im['activity_gt']
+
+        if gt == 'not started' or 'in between' in gt or gt == 'finished':
+            remove_images.append(gid)
+
+        """
+        aids = gid_to_aids[gid]
+        anns = ub.dict_subset(dset.anns, aids)
+
+        for aid, ann in anns.items():
+            conf = ann['confidence']
+            if conf < 0.4:
+                remove_anns.append(aid)
+        """
+
+    #print(f'removing {len(remove_anns)} annotations')       
+    #dset.remove_annotations(remove_anns)
+
+    print(f'removing {len(remove_images)} images (and associated annotations)')
+    dset.remove_images(remove_images)
+
+    # Save to a new dataset to adjust ids
+    new_dset = kwcoco.CocoDataset()
+    new_cats = [{"id": 1, "name": "tourniquet_tourniquet (step 1)"}, {"id": 2, "name": "tourniquet_tourniquet (step 2)"}, {"id": 3, "name": "tourniquet_tourniquet (step 3)"}, {"id": 4, "name": "tourniquet_tourniquet (step 4)"}, {"id": 5, "name": "tourniquet_tourniquet (step 5)"}, {"id": 6, "name": "tourniquet_tourniquet (step 6)"}, {"id": 7, "name": "tourniquet_tourniquet (step 7)"}, {"id": 8, "name": "tourniquet_tourniquet (step 8)"}, {"id": 9, "name": "tourniquet_label (step 1)"}, {"id": 10, "name": "tourniquet_label (step 2)"}, {"id": 11, "name": "tourniquet_label (step 3)"}, {"id": 12, "name": "tourniquet_label (step 4)"}, {"id": 13, "name": "tourniquet_label (step 5)"}, {"id": 14, "name": "tourniquet_label (step 6)"}, {"id": 15, "name": "tourniquet_label (step 7)"}, {"id": 16, "name": "tourniquet_label (step 8)"}, {"id": 17, "name": "tourniquet_windlass (step 1)"}, {"id": 18, "name": "tourniquet_windlass (step 2)"}, {"id": 19, "name": "tourniquet_windlass (step 3)"}, {"id": 20, "name": "tourniquet_windlass (step 4)"}, {"id": 21, "name": "tourniquet_windlass (step 5)"}, {"id": 22, "name": "tourniquet_windlass (step 6)"}, {"id": 23, "name": "tourniquet_windlass (step 7)"}, {"id": 24, "name": "tourniquet_windlass (step 8)"}, {"id": 25, "name": "tourniquet_pen (step 1)"}, {"id": 26, "name": "tourniquet_pen (step 2)"}, {"id": 27, "name": "tourniquet_pen (step 3)"}, {"id": 28, "name": "tourniquet_pen (step 4)"}, {"id": 29, "name": "tourniquet_pen (step 5)"}, {"id": 30, "name": "tourniquet_pen (step 6)"}, {"id": 31, "name": "tourniquet_pen (step 7)"}, {"id": 32, "name": "tourniquet_pen (step 8)"}, {"id": 33, "name": "hand (step 1)"}, {"id": 34, "name": "hand (step 2)"}, {"id": 35, "name": "hand (step 3)"}, {"id": 36, "name": "hand (step 4)"}, {"id": 37, "name": "hand (step 5)"}, {"id": 38, "name": "hand (step 6)"}, {"id": 39, "name": "hand (step 7)"}, {"id": 40, "name": "hand (step 8)"}]
+    for new_cat in new_cats:
+        new_dset.add_category(name=new_cat['name'], id=new_cat['id'])
+
+    for video_id, video in dset.index.videos.items():
+        new_dset.add_video(**video)
+
+    gid_to_aids = dset.index.gid_to_aids
+    gids = ub.argsort(ub.map_vals(len, gid_to_aids))
+
+    for gid in sorted(gids):
+        im = dset.imgs[gid]
+        new_im = im.copy()
+
+        aids = gid_to_aids[gid]
+        anns = ub.dict_subset(dset.anns, aids)
+
+        old_video = dset.index.videos[im['video_id']]['name']
+        new_video = new_dset.index.name_to_video[old_video]
+
+        del new_im['id']
+        new_im['video_id'] = new_video['id']
+        new_gid = new_dset.add_image(**new_im)
+
+        for aid, ann in anns.items():
+            old_cat = dset.cats[ann['category_id']]['name']
+            new_cat = new_dset.index.name_to_cat[old_cat]
+
+            new_ann = ann.copy()
+            del new_ann['id']
+            new_ann['category_id'] = new_cat['id']
+            new_ann['image_id'] = new_gid
+
+            new_dset.add_annotation(**new_ann)
+
+    new_dset.fpath = f'{experiment_name}_{stage}_{split}.mscoco.json'
+    new_dset.dump(new_dset.fpath, newlines=True)
+    print(f'Saved predictions to {new_dset.fpath}')
+
+def filter_kwcoco_conf_by_video(dset):
+    if type(dset) == str:
+        print(f'Loaded dset from file: {dset}')
+        dset = kwcoco.CocoDataset(dset)
+        print(dset)
+
+    gid_to_aids = dset.index.gid_to_aids
+    gids = ub.argsort(ub.map_vals(len, gid_to_aids))
+
+    remove_anns =[]
+    for gid in sorted(gids):
+        im = dset.imgs[gid]
+
+        aids = gid_to_aids[gid]
+        anns = ub.dict_subset(dset.anns, aids)
+
+        for aid, ann in anns.items():
+            conf = ann['confidence']
+            video_id = im['video_id']
+            video = dset.index.videos[video_id]['name']
+
+            if 'kitware' in video:
+                continue
+            else:
+                # filter the BBN videos by 0.4 conf 
+                if conf < 0.4:
+                    remove_anns.append(aid)
+
+    print(f'removing {len(remove_anns)} annotations')       
+    dset.remove_annotations(remove_anns)
+
+    dset.dump(dset.fpath, newlines=True)
+
 
 def main():
-    #coffee_root = '/Padlock_DT/Coffee'
-    #ros_bags_dir = f'{coffee_root}/coffee_recordings/extracted/'
-
-    #visualize_kwcoco(ros_bags_dir, 1, f'{ros_bags_dir}/coffee_contact_preds_with_background_all_objs_only_train.mscoco.json')
-
-
     ptg_root = '/data/ptg/medical/bbn/'
    
     kw = 'm2_all_data_cleaned_fixed_with_steps_results_train_activity.mscoco.json'
-    #kw = 'm2_with_lab_cleaned_fixed_data_with_inter_and_before_finished_steps_no_contact_aug_results_test.mscoco.json'
-    #kw = 'kitware_test_results_test.mscoco.json'
-    #kw = 'm2_with_lab_cleaned_fixed_data_with_inter_and_before_finished_steps_results_train_activity.mscoco.json'
 
     n = kw[:-12].split('_')
     name = '_'.join(n[:-1])
@@ -341,7 +333,7 @@ def main():
 
     stage = 'results'
     stage_dir = f'{ptg_root}/annotations/M2_Tourniquet/{stage}'
-    exp = 'm2_all_data_cleaned_fixed_with_steps'#'m2_with_lab_cleaned_fixed_data_with_inter_and_before_finished_steps'
+    exp = 'm2_all_data_cleaned_fixed_with_steps'
     if stage == 'stage1':
         save_dir = f'{stage_dir}/visualization_1/{split}'
     else:
@@ -350,7 +342,6 @@ def main():
     #save_dir = 'visualization'
     Path(save_dir).mkdir(parents=True, exist_ok=True)
     
-    #visualize_kwcoco(kw, save_dir)
     if stage == 'stage1':
         visualize_kwcoco(f'{stage_dir}/{kw}', save_dir)
     else:
