+++ angel_system/data/load_bbn_medical_data.py
@@ -13,7 +13,7 @@ import kwimage
 import pandas as pd
 import numpy as np
 
-from utils.data.dataloaders.common import time_from_name, activities_from_dive_csv
+from angel_system.data.common.load_data import activities_from_dive_csv
 
 os.environ['CUDA_VISIBLE_DEVICES'] = '0, 1, 2, 3'
 
@@ -228,26 +228,6 @@ def save_as_kwcoco(classes, data, save_fn='bbn-data.mscoco.json'):
 
     print_class_freq(dset)
 
-def print_class_freq(dset):
-    freq_per_class = dset.category_annotation_frequency()
-    stats = []
-
-    for cat in dset.cats.values():
-        freq = freq_per_class[cat['name']]
-        class_ = {
-            'id': cat['id'],
-            'name': cat['name'],
-            #'instances_count': freq,
-            #'def': '',
-            #'synonyms': [],
-            #'image_count': freq,
-            #'frequency': '',
-            #'synset': ''
-        }
-
-        stats.append(class_)
-
-    print(f'MC50_CATEGORIES = {stats}')
 
 def main():
     for split in ['train', 'test']:
